Sample: migrated the Reqnroll plugin to be a native Envelope consuming plugin (rather than writing then reading from a file).
