+++ Expressium.LivingDoc/Expressium.LivingDoc.csproj
@@ -38,7 +38,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="AngleSharp" Version="1.3.0" />
-		<PackageReference Include="Cucumber.Messages" Version="29.0.1" />
+		<PackageReference Include="Cucumber.Messages" Version="28.1.0" />
 	</ItemGroup>
 
 	<ItemGroup>
